Fix sign-comparison warnings

+++ a3/node.c
@@ -21,7 +21,7 @@ void run_node(int node_id, int ring_read_fd, int ring_write_fd, int stat_write_f
     while (1)
     {
         // Check for control messages from the parent
-        size_t bytes_read = read(ring_read_fd, &msg, sizeof(RingMessage));
+        int bytes_read = read(ring_read_fd, &msg, sizeof(RingMessage));
         if (bytes_read == -1)
         {
             if (errno == EINTR)
@@ -38,7 +38,7 @@ void run_node(int node_id, int ring_read_fd, int ring_write_fd, int stat_write_f
         {
             if (msg.type == MSG_SHUTDOWN)
             {
-                size_t pass_shutdown_bytes = write(ring_write_fd, &msg, sizeof(RingMessage));
+                int pass_shutdown_bytes = write(ring_write_fd, &msg, sizeof(RingMessage));
                 printf("Node %d received shutdown message and is passing it to the next node.\n", node_id);
                 if (pass_shutdown_bytes == -1)
                 {
@@ -59,7 +59,7 @@ void run_node(int node_id, int ring_read_fd, int ring_write_fd, int stat_write_f
                 printf("Node %d received a token.\n", node_id);
                 // Pass the token to the next node
                 msg.hop_count++; // Increment hop count for the token
-                size_t bytes_written = write(ring_write_fd, &msg, sizeof(RingMessage));
+                int bytes_written = write(ring_write_fd, &msg, sizeof(RingMessage));
                 if (bytes_written == -1)
                 {
                     printf("Node %d failed to pass the token.\n", node_id);
@@ -74,7 +74,7 @@ void run_node(int node_id, int ring_read_fd, int ring_write_fd, int stat_write_f
             {
                 // If it is a task message for a different node, we should still pass it along
                 msg.hop_count++; // Increment hop count for the message
-                size_t bytes_written = write(ring_write_fd, &msg, sizeof(RingMessage));
+                int bytes_written = write(ring_write_fd, &msg, sizeof(RingMessage));
                 if (bytes_written == -1)
                 {
                     printf("Node %d failed to pass the task message.\n", node_id);
@@ -105,7 +105,7 @@ void run_node(int node_id, int ring_read_fd, int ring_write_fd, int stat_write_f
 
                 // Send the report back to the parent
                 RingMessage report_msg = make_report_msg(node_id, msg.task_id, msg.sequence_num, msg.status, result);
-                size_t stat_bytes_written = write(stat_write_fd, &report_msg, sizeof(RingMessage));
+                int stat_bytes_written = write(stat_write_fd, &report_msg, sizeof(RingMessage));
                 if (stat_bytes_written == -1)
                 {
                     printf("Node %d failed to send the report to the parent.\n", node_id);
